+++ src/services/contentful.ts
@@ -5,7 +5,7 @@ import {
   GenericItemProps,
 } from 'types/services-types';
 
-import fetchWithCache from 'services/fetch/fetch';
+import fetchWithCache from 'services/fetch';
 
 const { REACT_APP_CONTENTFUL_SPACE_ID } = process.env;
 const { REACT_APP_CONTENTFUL_TOKEN } = process.env;
